+++ useOnClickRouteLink.test.mjs
@@ -76,11 +76,11 @@ Deno.test("`useOnClickRouteLink` memoization.", () => {
 
 Deno.test("`useOnClickRouteLink` in a DOM environment.", async () => {
   const abortController = new AbortController();
-  const projectFileServer = await serveProjectFiles(abortController.signal);
+  const projectFileServer = serveProjectFiles(abortController.signal);
 
   try {
     const projectFilesOriginUrl = new URL(
-      `http://localhost:${projectFileServer.port}`,
+      `http://localhost:${projectFileServer.addr.port}`,
     );
     const clientImportMap = await readImportMapFile(
       new URL("importMap.json", import.meta.url),
@@ -221,6 +221,6 @@ Deno.test("`useOnClickRouteLink` in a DOM environment.", async () => {
     }
   } finally {
     abortController.abort();
-    await projectFileServer.close;
+    await projectFileServer.finished;
   }
 });