+++ changelog.md
@@ -26,6 +26,7 @@
 - Added in project and readme scripts the CLI `deno` flag `--allow-import`, as
   required by Deno v2.
 - Modernized the project type check script.
+- Modernized the test helper function `serveProjectFiles`.
 - Added a project `.zed/settings.json` file to make is easier to contribute
   using the Zed editor.
 - Migrated from